Render 12 months with according days
