+++ client/main.css
@@ -15,6 +15,7 @@ body {
   width: 90vw;
   height: 100%;
   margin: 0 auto;
+  padding-bottom: 3rem;
 }
 
 .app h1,
@@ -32,7 +33,7 @@ h3 {
   flex-direction: row;
   justify-content: space-around;
   align-items: center;
-  width: 80vw;
+  width: 90vw;
   text-align: center;
   margin: 0 auto;
 }
@@ -42,9 +43,8 @@ h3 {
   height: 2rem;
   border: thin solid black;
   outline: none;
-  margin: 0.2rem;
   border-radius: 5px;
-  padding: 0;
+  margin: 0 auto;
 }
 
 .square:hover {
@@ -53,18 +53,23 @@ h3 {
 }
 
 .table {
-  display: flex;
+  display: grid;
+  grid-template-columns: repeat(4, 1fr);
+  gap: 3rem;
   width: 90vw;
-  justify-content: center;
 }
 
 .container {
   border: solid thin grey;
-  border-radius: 10px;
-  padding: 1rem;
-  width: 68vw;
-  margin-top: 1.5rem;
   background-color: white;
+  padding: 1rem;
+  width: 16vw;
+  margin: 0 auto;
+}
+
+.container p {
+  text-align: center;
+  margin-top: 0;
 }
 
 .selectors {
@@ -79,4 +84,6 @@ h3 {
   display: flex;
   flex-direction: row;
   align-items: center;
+  justify-content: space-between;
+  width: 20%;
 }
